test(device-mode): add late-subscriber regression tests for event filtering
- merge standalone late-subscriber test into EventFilteringPluginTest
- cover both denylist and allowlist warm-launch + identical-refresh cases

Scanned-by: gitleaks 8.29.0
